разбил на модули: util.js и data.js как точка входа. main.js пока пустой
